move grouping config constants to `server.py`
