+++ tests/sentry/grouping/__init__.py
@@ -10,6 +10,7 @@ import pytest
 from django.utils.functional import cached_property
 
 from sentry import eventstore
+from sentry.conf.server import DEFAULT_GROUPING_CONFIG
 from sentry.event_manager import EventManager, get_event_type, materialize_metadata
 from sentry.eventstore.models import Event
 from sentry.grouping.api import (
@@ -24,7 +25,6 @@ from sentry.grouping.fingerprinting import FingerprintingRules
 from sentry.grouping.strategies.configurations import CONFIGURATIONS
 from sentry.grouping.variants import BaseVariant
 from sentry.models.project import Project
-from sentry.projectoptions.defaults import DEFAULT_GROUPING_CONFIG
 from sentry.stacktraces.processing import normalize_stacktraces_for_grouping
 from sentry.testutils.helpers.eventprocessing import save_new_event
 from sentry.utils import json
