+++ tests/sentry/tasks/test_reprocessing2.py
@@ -8,6 +8,7 @@ import pytest
 
 from sentry import eventstore
 from sentry.attachments import attachment_cache
+from sentry.conf.server import DEFAULT_GROUPING_CONFIG
 from sentry.event_manager import EventManager
 from sentry.eventstore.models import Event
 from sentry.eventstore.processing import event_processing_store
@@ -20,7 +21,6 @@ from sentry.models.groupassignee import GroupAssignee
 from sentry.models.groupredirect import GroupRedirect
 from sentry.models.userreport import UserReport
 from sentry.plugins.base.v2 import Plugin2
-from sentry.projectoptions.defaults import DEFAULT_GROUPING_CONFIG
 from sentry.reprocessing2 import is_group_finished
 from sentry.tasks.reprocessing2 import finish_reprocessing, reprocess_group
 from sentry.tasks.store import preprocess_event
